Use flexbox to generalize top-level layout
This replaces the absolute positioning of the header and footer that was
used to ensure that they were at the top and bottom of the viewport,
even if the page content didn't reach all the way to the bottom of the
viewport. Using flexbox has the advantage of not hard-coding heights for
the header and footer.