Prevent Cloudflare from obfuscating email address
The obfuscation is probably helping a bit with spam reduction, but I
care more about not needing to load extra JavaScript and making the site
totally functional in minimal browsers.

In order to keep the HTML comments, the `--minify` flag can't be used
when building, so I switched to minifying the CSS only.