Give code creations a real page
I want these new pages to be the canonical home for my code projects.
Hosting platforms like SourceHut and GitHub are useful for the services
they provide, and I'll continue to use them for code hosting, mailing
lists, etc., but I want to own the "home page" for all of my projects.

As part of this change, update the last updated blurb to be clear that
it indicates when the page was last updated, not the project itself.